Create models folder and move building_classifier files into it
